Add files via upload
This is the 3rd version of the original program. 
Due to fstream's inability to write null pointer arrays into a file, I have removed the stack feature from the class user.
Due to this, only one book can be rented by an user. 
However, this program is fully functional now.